Remove startlingmo
