--- startlingmo/config-startlingmo.h.cmake
@@ -1,9 +1,0 @@
-#pragma once
-
-#define CMAKE_INSTALL_FULL_BINDIR "@CMAKE_INSTALL_FULL_BINDIR@"
-#define KDE_INSTALL_FULL_DATAROOTDIR "@KDE_INSTALL_FULL_DATAROOTDIR@"
-#define CMAKE_INSTALL_FULL_LIBEXECDIR "@CMAKE_INSTALL_FULL_LIBEXECDIR@"
-#define CMAKE_INSTALL_FULL_LIBEXECDIR_KF6 "@CMAKE_INSTALL_FULL_LIBEXECDIR_KF6@"
-#define KWIN_WAYLAND_BIN_PATH "@KWIN_WAYLAND_BIN_PATH@"
-
-#define KWIN_BIN "${KWIN_BIN}"
