--- startlingmo/lingmo-session/CMakeLists.txt
@@ -1,46 +1,0 @@
-project(lingmo_session)
-
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED 17)
-
-set(TARGET lingmo_session)
-
-set(SOURCES
-        application.cpp
-        main.cpp
-        process.cpp
-        processmanager.cpp
-        networkproxymanager.cpp
-
-        powermanager/power.cpp
-        powermanager/powerproviders.cpp
-)
-
-qt_add_dbus_adaptor(DBUS_SOURCES
-                     com.lingmo.Session.xml
-                     application.h Application
-                     sessionadaptor SessionAdaptor)
-# set_source_files_properties(${DBUS_SOURCES} PROPERTIES SKIP_AUTOGEN ON)
-
-find_package(KF5WindowSystem)
-find_package(Threads)
-
-ecm_qt_declare_logging_category(SOURCES  HEADER debug.h IDENTIFIER LINGMO_SESSION_D CATEGORY_NAME org.lingmo.session)
-
-add_executable(${TARGET} ${SOURCES} ${DBUS_SOURCES})
-target_link_libraries(${TARGET}
-    PRIVATE
-        startlingmo
-    PUBLIC
-        Qt5::Core
-        Qt5::Gui
-        Qt5::Widgets
-        Qt5::Quick
-        Qt5::DBus
-        Qt5::X11Extras
-        KF5::WindowSystem
-        ${CMAKE_THREAD_LIBS_INIT}
-)
-target_include_directories(${TARGET} PRIVATE ${CMAKE_SOURCE_DIR}/startlingmo)
-
-install(TARGETS ${TARGET} DESTINATION ${CMAKE_INSTALL_BINDIR})
