--- startlingmo/lingmo-session/com.lingmo.Session.xml
@@ -1,36 +1,0 @@
-<!DOCTYPE node PUBLIC "-//freedesktop//DTD D-BUS Object Introspection 1.0//EN" "http://www.freedesktop.org/standards/dbus/1.0/introspect.dtd">
-<node>
-  <interface name="com.lingmo.Session">
-    <method name="logout">
-      <annotation name="org.freedesktop.DBus.Method.NoReply" value="true"/>
-    </method>
-    <method name="reboot">
-      <annotation name="org.freedesktop.DBus.Method.NoReply" value="true"/>
-    </method>
-    <method name="powerOff">
-      <annotation name="org.freedesktop.DBus.Method.NoReply" value="true"/>
-    </method>
-    <method name="suspend">
-      <annotation name="org.freedesktop.DBus.Method.NoReply" value="true"/>
-    </method>
-    <method name="startDesktopProcess">
-      <annotation name="org.freedesktop.DBus.Method.NoReply" value="true"/>
-    </method>
-    <method name="updateNetworkProxy">
-      <annotation name="org.freedesktop.DBus.Method.NoReply" value="true"/>
-    </method>
-    <method name="launch">
-      <arg name="exec" type="s" direction="in"/>
-      <arg name="args" type="as" direction="in"/>
-    </method>
-    <method name="launch">
-      <arg name="exec" type="s" direction="in"/>
-      <arg name="workingDirectory" type="s" direction="in"/>
-      <arg name="args" type="as" direction="in"/>
-    </method>
-    <method name="updateLaunchEnv">
-      <arg name="key" type="s" direction="in"/>
-       <arg name="value" type="s" direction="in"/>
-    </method>
-  </interface>
-</node>
