--- startlingmo/lingmo-wayland-session.desktop
@@ -1,6 +1,0 @@
-[Desktop Entry]
-Type=Application
-Exec=startlingmo-wayland
-Name=Lingmo Desktop (Wayland Experimental)
-Keywords=session
-Comment=session
