--- startlingmo/lingmo-xorg-session.desktop
@@ -1,7 +1,0 @@
-[Desktop Entry]
-Type=Application
-Exec=startlingmo-x11
-TryExec=startlingmo-x11
-Name=Lingmo Desktop (Xorg Session)
-Keywords=session
-Comment=session
