--- startlingmo/wayland_wrapper/lib/xauthority.h
@@ -1,14 +1,0 @@
-/*
-    KWin - the KDE window manager
-    This file is part of the KDE project.
-
-    SPDX-FileCopyrightText: 2021 David Edmundson <davidedmundson@kde.org>
-
-    SPDX-License-Identifier: GPL-2.0-or-later
-*/
-
-#pragma once
-
-class QTemporaryFile;
-
-bool generateXauthorityFile(int display, QTemporaryFile *authorityFile);
